Update patch set 7
Patch Set 7:

> Any reason the weather_ strings numbering was changed compared to
 > LockClock? This breaks the ability to import existing translations.

The reason behind this decision was that there are 3 duplicates in the weather codes. 

11	showers
12	showers


38	scattered thunderstorms
39	scattered thunderstorms


41	heavy snow
43	heavy snow


The weather codes in the new WeatherContract in our sdk is a port from the codes defined in LockClock , but I removed those 3 duplicates

public static final class WeatherCode {
   ...
   public final static int SHOWERS = 11;
   ...
   public final static int SCATTERED_THUNDERSTORMS = 37;
   ...
   public final static int HEAVY_SNOW = 39;
   ...
}

We could bring those codes back into the WeatherContract (and consequently to the weather provider) in order to keep a one to one mapping with the codes in LockClock, but we'd still need a different constant name (and therefore a different string) for those codes anyway (for example, 11 could be rain showers and 12 could be heavy showers, etc) so I'm not sure that would help.

Patch-set: 7